[IPLUG] added IPlugBase::OnHostIdentified(), added VST3 host detection,
to be overridden by your plugin class, and can be used to do something once the host and version has been identified (which may happen after the plugin's constructor has been called for audiounits). HostSpecificInit() is not pure virtual and is removed from API wrappers that don't need it (for now).