BREAKING: Modify IPopupMenu behaviour in order to supportdrawing contextual menus with IGraphics
An IPopupMenu cannot be a temporary - it should be a member variable. Previously we relied on the (blocking) behaviour of platform pop-up menus on mac os and windows in order to get results from a pop-up menu synchronously - which meant temporary IPopupmenus were viable. This no longer happens. IControl now has a method OnPopupMenuSelection() which will fire on completion of the pop-up menu. Alternatively you can use a lamda function of IPopupMenu