IMouseMod required before defined, when building AAX
IGraphicsStruct.h includes IGraphicsUtilities.h, which uses IMouseMod, before defining IMouseMod. This should be fixed more elegantly. As a quick fix, I commented out the section of code in IGraphicsUtilities.h, to leave a note, and copied it to the end of IGraphicsStruct.h, a place it doesn't belong logically.