Temporarily patch NanoVG for crash on app-quit
We are behind the NanoVG tip, which has some issues (e.g. https://github.com/iPlug2/iPlug2/issues/152). This  patch should  in theory fix an occasional crash on quit  with the current version in the IPlug source tree,  however,  will need to look at it again when  updating.

 occasionally I get a crush on quit and it appears to be an assertion:
BUG IN CLIENT OF LIBDISPATCH: Semaphore object deallocated while in use

A workaround is described here:
https://lists.apple.com/archives/cocoa-dev/2014/Apr/msg00484.html