macOS: Fix window dims when resizing notch macs
When using the IGraphics corner resizer on M1 Pro/Max "notch" macs, the window size ended up wrong, since we were using the height of the macOS menu bar to adjust the client rect. On macs with no notch, this worked, but on macs with a notch the height of the menu bar is different to that of a windows title bar.