iPlug2 WebView Library is MIT licensed
Whilst the rest of iPlug2 is Zlib-style the contents of Extras/WebView is MIT licensed.
I am not going to rewrite git-history, but I would appreciate if you can treat it
as if it is has always been this way. This just requires you to credit my work if you use
it by including it in your documentation or any derivative works. I'd hope you might do this anyway.