Update ITextEntryControl.cpp to support label entry
This is an initial suggestion of how you could go about entering the label as a part of the text entry. For simple examples this already works, but there are some outstanding issues:

1.) It's case sensitive, and probably shouldn't be
2.) You cannot use this with advanced entry, like entries where the label changes type depending on the value
3.) It's limited to a double, it could be extended for all

Just throwing this out there to show that it's both possible, and probably not THAT hard.

What do people want to see to consider this complete?