Do not write IXML BWF_TIME_REFERENCE_LOW/BWF_TIME_REFERENCE_HIGH with 0 values
the previous commit and this are trying to get media explorer metadata edits to happen
completely in place if possible. the one remaining issue is that we
can't split up the various RIFF chunks to match the previous gaps. so
it's fruitless, but trying has spotted these differences between writing
and re-writing metadata -- from c18ef808